Add interview types and application status to profile page
